pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "paper-analyzer"
version = "0.1.0"
description = "本地运行的论文分析助手"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26,<3",
    "sentence-transformers>=3,<6",
    "PyMuPDF>=1.24,<2",
    "pytesseract>=0.3.10,<1",
    "pdf2image>=1.16,<2",
    "openai>=1.0,<3",
    "python-dotenv>=1.0,<2",
    "beautifulsoup4>=4.12,<5",
    "requests>=2.31,<3",
]

[project.optional-dependencies]
web = ["streamlit>=1.30,<2"]
dev = ["pytest>=8,<10"]

[project.scripts]
paper-analyzer = "main:main"

[tool.setuptools.packages.find]
include = ["paper_analyzer*", "pipeline*"]
